Revised the Windows build system in the following manner:
   1. Introduced a peace_library project to build a static library just like
      it is done under Windows.
   2. Revied the peace project to utilize the library
   3. Introduced a new project for building examples under Windows as per
      Karro's request.

Fixed a couple of minor type mismatch warnings that vc9 was unhappy about
(size_t to int conversions)

Added a check to handle a sentinel EST file name called "<none>" that can
be set by programs to specify a dummy EST file to simply suppress error that
a EST file has not been specified. This is used from UsingPeace.cpp program.

Added a simple example code to UsingPeace.cpp to illustrate how parameters
can be set for d2.

+++ peace/src/CLU.cpp
@@ -187,7 +187,7 @@ CLU::createCLUHashMap(int* &hashMap, const char *sequence) {
     // hash for subsequent words can be easily derived from it using
     // bit-wise operations as the word simply slides acorss the
     // sequence.
-    const int SeqLen  = strlen(sequence) - wordSize;
+    const int SeqLen  = (int) strlen(sequence) - wordSize;
     // Compute bit mask that will retain only the bits corresponding
     // to a given word size.  Each entry in a word takes up 2 bits and
     // that is why the following formula involves a 2.
@@ -289,7 +289,7 @@ CLU::getSimilarity(const int* const hashMap,
     ASSERT ( (int) prevHashValues.size() == frameSize );
     // Now using the hash and cdIndex values for first frame compute
     // the values for other frames by suitably adjusting the values.
-    const int SeqLen = strlen(sequence) - frameSize;
+    const int SeqLen = (int) strlen(sequence) - frameSize;
     for(int i = frameSize; (i < SeqLen); i++) {
         // Ensure cdIndex never exceeds the number of cateogries.
         cdIndex = (cdIndex >= (unsigned) frameSize) ? (frameSize - 1) : cdIndex;
